feat(pmtiles): make Offline Basemap Extract a floating panel with draw-bbox
Replace the modal BasemapExtractDialog with BasemapExtractPanel, a non-modal
floating draggable panel mounted over the map (like RasterSubsetPanel), so the
map stays interactive while extracting.

- Draw on map: rubber-band a bbox on the map (dragPan/boxZoom suspended for the
  drag, Esc/blur cancel), with a dashed SVG overlay tracking the box under
  pan/zoom. Keeps "Use view" and the editable coordinate fields.
- Mounting moves from a TopToolbar <Dialog> to DesktopShell's MapGrid; the Add
  Data menu opens it via a new onOpenBasemapExtract toolbar prop.
- All extract logic (URL, zoom range, progress bar, size confirmation, save,
  add-layer) is unchanged. New i18n keys: drawBbox/drawing/drawHint.

Verified in a real browser: floating draggable panel, drew a bbox (overlay +
coords), and extracted Florence z0-15 from the live Protomaps planet build
(saved + added to the map), in both light and dark themes.
