+++ apps/geolibre-desktop/src/i18n/locales/en.json
@@ -1184,7 +1184,10 @@
     "urlPlaceholder": "https://data.source.coop/…/planet.pmtiles",
     "urlHint": "The host must allow cross-origin HTTP range requests (source.coop does; the Protomaps build server does not).",
     "area": "Area",
-    "useView": "Use current view",
+    "useView": "Use view",
+    "drawBbox": "Draw on map",
+    "drawing": "Drawing…",
+    "drawHint": "Drag a box on the map to set the extract area.",
     "north": "North",
     "south": "South",
     "west": "West",
